cart sorting

+++ frontend/src/features/cartSlice.js
@@ -11,8 +11,19 @@ const cartSlice = createSlice({
     initialState,
     reducers: {
         addToCart: (state, action) => {
-            state.items.push(action.payload);
-            localStorage.setItem("cartItems",JSON.stringify(state.items));
+            const exists = state.items.find((item) => item._id === action.payload._id);
+            if(!exists){
+                state.items.push(
+                    {
+                        ...action.payload,
+                        addedAt: new Date().toISOString()
+                    }
+                );
+                localStorage.setItem("cartItems",JSON.stringify(state.items));
+            } else {
+                alert("Item already in cart");
+            }
+
         },
         removeFromCart: (state, action) => {
         state.items = state.items.filter((item) => item._id !== action.payload);
